% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-location.R
\name{find_location}
\alias{find_location}
\title{Find DC Location}
\usage{
find_location(string)
}
\arguments{
\item{string}{address string}
}
\value{
tibble
}
\description{
Uses DC's \href{https://dcatlas.dcgis.dc.gov/mar/}{Master Address Repository (MAR)} to find a DC-based
location. It returns a table of all possible matching
addresses based on the provided string.
}
\examples{
find_location(string = "1600 Pennsylvania Ave NW")
}
