% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidymar-package.R
\docType{package}
\name{tidymar}
\alias{tidymar}
\alias{tidymar-package}
\title{Query DC addresses, intersections, blocks, and place names}
\description{
District of Columbia's \href{https://dcatlas.dcgis.dc.gov/mar/}{Master Address Repository (MAR)} provides a \href{https://opendata.dc.gov/pages/mar-webservices}{web service}
with geocoding and address verification operations. This package takes
advantage of the MAR's addressing webservice.
}
\author{
Hersh Gupta
}
