% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-location.R
\name{find_location}
\alias{find_location}
\title{Find DC Location}
\usage{
find_location(address)
}
\arguments{
\item{address}{address string or vector of addresses}
}
\value{
tibble
}
\description{
Uses DC's \href{https://dcatlas.dcgis.dc.gov/mar/}{Master Address Repository (MAR)} to find a DC-based
location. It returns a table of all possible matching
addresses based on the provided string.
}
\examples{
\dontrun{
find_location(address = '1600 Pennsylvania Ave NW')
}

}
